[TASK] sync Mexec Proxy now declared in separate module
